+++ src/bp/ml/ml-thread-index.ts
@@ -8,7 +8,7 @@ import * as sdk from 'botpress/sdk'
 import { parentPort } from 'worker_threads'
 
 import { Trainer as CrfTrainer } from './crf'
-import { Message } from './ml-worker-pool'
+import { Message } from './ml-thread-pool'
 import { Trainer as SvmTrainer } from './svm'
 
 // Debugging currently not possible in this file and beyond...
